# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kanban API",
        "description": "Backend API for Kanban App — RESTful service managing users, boards, tasks, and authentication for daily task management.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/boards/active": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Get the currently active board",
                "operationId": "f61dd3d2537c4a252556aaf62c879417",
                "responses": {
                    "200": {
                        "description": "Active board retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "board": {
                                                    "$ref": "#/components/schemas/Board"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active board found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No active boards found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    },
                    "500": {
                        "description": "Failed to get an active Board",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/boards/{id}/activate": {
            "patch": {
                "tags": [
                    "Boards"
                ],
                "summary": "Sets a board as active",
                "operationId": "d0b729bdc9ab5b94be5d71750b873d89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board set as active",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Board set as active."
                                        },
                                        "data": {
                                            "properties": {
                                                "board": {
                                                    "$ref": "#/components/schemas/Board"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Board not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Board] 5"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    },
                    "500": {
                        "description": "Failed to set board as active",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Board": {
                "title": "Board",
                "description": "Represents a board, including its columns and archived status.",
                "properties": {
                    "id": {
                        "description": "Unique identifier of the board.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the board. Must be unique per user.",
                        "type": "string",
                        "example": "My Project Board"
                    },
                    "is_archived": {
                        "description": "Indicates whether the board is archived (true) or active (false).",
                        "type": "boolean",
                        "example": false
                    },
                    "columns": {
                        "description": "List of columns that belong to this board.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Column"
                        }
                    }
                },
                "type": "object"
            },
            "Column": {
                "required": [
                    "name",
                    "order"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "To Do"
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Task"
                        }
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    },
                    "error": {
                        "type": "string",
                        "example": "Error details"
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "new_column_id": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The selected new_column_id is invalid."
                                }
                            },
                            "new_order": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The new_order must be at least 0."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}